rcu: Prevent early-boot RCU callbacks from splatting
Currently, a call_rcu() that precedes rcu_init() will splat due to the
callback lists not having yet been initialized.  This commit causes the
first such callback to initialize the boot CPU's RCU callback list.

Note that this commit does not change rcu_init()-time initialization,
which means that the callback will be discarded at rcu_init() time.
Fixing this is the job of later commits.

Change-Id: I69bfc1038fb89c6e87d2a56cda20c6067308c802
Signed-off-by: Paul E. McKenney <paulmck@linux.vnet.ibm.com>
Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>
Signed-off-by: Ad!thya R <gh0strider.2k18.reborn@gmail.com>